feat: added basic snapshot tests for step wizard
